pyproject.toml
[project]
name = "parameter_variability"
version = "0.1.0"
description = "Parameter variability in SBML models"
readme = "README.md"
requires-python = ">=3.13"
authors = [
	{name="Matthias König", email="konigmatt@googlemail.com"},
	{name="Antonio Alvarez", email="antonio.alvarez@student.hu-berlin.de"},
]
maintainers = [
	{name="Matthias König", email="konigmatt@googlemail.com"}
]
classifiers = [
	"Development Status :: 4 - Beta",
	"Intended Audience :: Science/Research",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
   "Programming Language :: Python :: 3.13",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Scientific/Engineering",
	"Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = [
	"SBML",
	"variability",
]
dependencies = [

    "sbmlutils>=0.9.6",
    "libroadrunner>=2.9.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pydantic",
    "pydantic-yaml",

    "matplotlib>=3.10.0",

    # optimization
    "pypesto>=0.5.8",
    "pypesto[fides]",
    "fides>=0.8.0",
    "petab>=0.8.1",
    "amici>=0.34.2",

    "arviz>=0.20.0",
    "rich>=13.9.4",
    "xarray>=2025.1.2",
    "graphviz",

    "pytest",
]

[project.optional-dependencies]
dev = [
    "bump-my-version>=1.2.6",
	"ruff>=0.14.11",
	"pre-commit>=4.5.1",
	"mypy>=1.19.1",
    "tox>=4.34.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[project_urls]
Homepage = "https://github.com/matthiaskoenig/parameter-variability"
Repository = "https://github.com/matthiaskoenig/parameter-variability"
Documentation = "https://github.com/matthiaskoenig/parameter-variability"
Issues = "https://github.com/matthiaskoenig/parameter-variability/issues"
